rename test-util to testutil_shrinkwrap
